Refactor: Moved useLocallyPersistedReducer out into a utils function
