+++ TODOS
@@ -1,4 +1,3 @@
-- Update Dashboard/Player reset functionality to clear just the key.
 - Persist Dark mode in local storage
 - Build a Scoreboard View
 - Allow Persist the Scoreboard for that user.
